Change project structure and split code into several python files
