+++ src/pyprobs/__init__.py
@@ -1,0 +1,5 @@
+from pyprobs._version import __version__
+
+from pyprobs.probability import Probability
+
+__all__ = ['Probability']
